add : 공용컴포넌트 수정 - develop 폴더 위치 변경에 맞추어 기존의 공용컴포넌트 폴더 위치 변경 및 경로 수정
