+++ libs/core/src/components/button/AskSelectButton.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
-import { customColor, Typography } from '@street-vendor/core';
+import { customColor } from '../../constants';
+import { Typography } from '../Typography';
 
 interface ButtonProps {
   content: string;
