+++ libs/core/src/components/form/AskOption.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
-import { customColor, Typography } from '@street-vendor/core';
+import { customColor } from '../../constants';
+import { Typography } from '../Typography';
 
 interface OptionProps {
   value: string;
