+++ libs/core/src/components/form/AskTextarea.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
-import { customColor, Typography } from '@street-vendor/core';
+import { customColor } from '../../constants';
+import { Typography } from '../Typography';
 import {
   FieldErrorsImpl,
   FieldValues,
